+++ src/lib/output/themes/default/assets/typedoc/Component.ts
@@ -8,12 +8,13 @@ export interface IComponentOptions {
 /**
  * TypeDoc component class.
  */
-export class Component {
-    protected el: HTMLElement;
+export class Component<E extends HTMLElement = HTMLElement> {
+    protected el: E;
     protected app: Application;
 
     constructor(options: IComponentOptions) {
-        this.el = options.el;
+        this.el = options.el as E;
         this.app = options.app;
+        console.log("Created", this, "for", options);
     }
 }