+++ src/lib/output/themes/default/partials/member.declaration.tsx
@@ -1,34 +1,11 @@
-import type { DeclarationReflection, ReflectionType } from "../../../../models/index.js";
-import { JSX, Raw } from "../../../../utils/index.js";
+import type { DeclarationReflection } from "../../../../models/index.js";
+import { JSX } from "../../../../utils/index.js";
 import { getKindClass, hasTypeParameters, renderTypeParametersSignature, wbr } from "../../lib.js";
 import type { DefaultThemeRenderContext } from "../DefaultThemeRenderContext.js";
 
-function renderingTypeDeclarationIsUseful(declaration: DeclarationReflection): boolean {
-    if (declaration.hasComment()) return true;
-    if (declaration.children?.some(renderingTypeDeclarationIsUseful)) return true;
-    if (declaration.type?.type === "reflection" && renderingTypeDeclarationIsUseful(declaration.type.declaration)) {
-        return true;
-    }
-
-    return declaration.getAllSignatures().some((sig) => {
-        return sig.hasComment() || sig.parameters?.some((p) => p.hasComment());
-    });
-}
+void JSX; // TS is confused and thinks this is unused
 
 export function memberDeclaration(context: DefaultThemeRenderContext, props: DeclarationReflection) {
-    function renderTypeDeclaration(type: ReflectionType) {
-        if (renderingTypeDeclarationIsUseful(type.declaration)) {
-            return (
-                <div class="tsd-type-declaration">
-                    <h4>{context.i18n.theme_type_declaration()}</h4>
-                    {context.parameter(type.declaration)}
-                </div>
-            );
-        }
-    }
-
-    const visitor = { reflection: renderTypeDeclaration };
-
     return (
         <>
             <div class="tsd-signature">
@@ -54,29 +31,7 @@ export function memberDeclaration(context: DefaultThemeRenderContext, props: Dec
 
             {hasTypeParameters(props) && context.typeParameters(props.typeParameters)}
 
-            {props.type?.visit<JSX.Children>({
-                reflection: renderTypeDeclaration,
-                array: (arr) => arr.elementType.visit(visitor),
-                intersection: (int) => int.types.map((t) => t.visit(visitor)),
-                union: (union) => {
-                    if (union.elementSummaries) {
-                        const result: JSX.Children = [];
-                        for (let i = 0; i < union.types.length; ++i) {
-                            result.push(
-                                <li>
-                                    {context.type(union.types[i])}
-                                    <Raw html={context.markdown(union.elementSummaries[i])} />
-                                    {union.types[i].visit(visitor)}
-                                </li>,
-                            );
-                        }
-                        return <ul>{result}</ul>;
-                    }
-                    return union.types.map((t) => t.visit(visitor));
-                },
-                reference: (ref) => ref.typeArguments?.map((t) => t.visit(visitor)),
-                tuple: (ref) => ref.elements.map((t) => t.visit(visitor)),
-            })}
+            {props.type && context.typeDeclaration(props.type)}
 
             {context.commentTags(props)}